requirements.txt
mcp
openai
requests
google-adk